cleaned up folders tweaked backend
