--- backend/db/schema.sql
@@ -1,2 +1,0 @@
-DROP DATABASE IF EXISTS projectManagement_db;
-CREATE DATABASE projectManagement_db;
