--- backend/db/seeds.sql
@@ -1,18 +1,0 @@
--- Insert multiple project items --
-INSERT INTO projects (id, project_name, active)
-VALUES
-    ( 1, "Quartz", true),
-    ( 2, "Jet", true),
-    ( 3, "Amethyst", false)
-
--- Insert multiple user items --
-INSERT INTO users (id, first_name, last_name)
-VALUES
-    ( 1, "Andrew", "Edwards", true, true),
-    ( 2, "Itzel", "Cole", false, true),
-    ( 3, "Joel", "Prentiss", true, false);
-       
-
-
-       
-
