--- backend/model/index.js
@@ -1,13 +1,0 @@
-const User = require('./userModel');
-const Project = require('./projectModel');
-
-User.hasMany(Project, {
-    foreignKey: 'user_id',
-    onDelete: 'CASCADE'
-});
-
-Project.belongsTo(User, {
-    foreignKey: 'user_id'
-});
-
-module.exports = { User, Project };
