--- backend/model/projectModel.js
@@ -1,60 +1,0 @@
-const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../config/connection');
-
-class Project extends Model { }
-
-Project.init(
-    {
-        id: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
-            primaryKey: true,
-            autoIncrement: true,
-        },
-        name: {
-            type: DataTypes.STRING,
-            allowNull: false,
-        },
-        budget: {
-            type: DataTypes.FLOAT,
-            allowNull: true,
-        },
-        status: {
-            type: DataTypes.STRING,
-            allowNull: true,
-        },
-        deadline: {
-            type: DataTypes.DATEONLY,
-            allowNull: true,
-        },
-        priority: {
-            type: DataTypes.STRING,
-            allowNull: true,
-        },
-        description: {
-            type: DataTypes.STRING,
-        },
-        date_created: {
-            type: DataTypes.DATE,
-            allowNull: false,
-            defaultValue: DataTypes.NOW,
-        },
-
-        user_id: {
-            type: DataTypes.INTEGER,
-            references: {
-                model: 'user',
-                key: 'id',
-            },
-        },
-    },
-    {
-        sequelize,
-        timestamps: false,
-        freezeTableName: true,
-        underscored: true,
-        modelName: 'project',
-    }
-);
-
-module.exports = Project;
