--- backend/seeds/projectData.json
@@ -1,27 +1,0 @@
-[
-    {
-      "name": "Project 2",
-      "description": "Add handlebars",
-      "budget": 0,
-      "priority": "High",
-      "status": "In progress",
-      "deadline": "2022-08-08T03:39:48.678Z"
-    },
-    {
-      "name": "Project 2",
-      "description": "Add routes",
-      "budget": 0,
-      "priority": "High",
-      "status": "In progress",
-      "deadline": "2022-08-08T03:39:48.678Z"
-    },
-    {
-      "name": "Project 2",
-      "description": "Add routes",
-      "budget": 0,
-      "priority": "High",
-      "status": "In progress",
-      "deadline": "2022-08-08T03:39:48.678Z"
-    }
-  ]
-
