--- backend/seeds/userData.json
@@ -1,23 +1,0 @@
-[
-    {
-      "name": "Brian",
-      "email": "brian@quartz.com",
-      "password": "password"
-    },
-    {
-      "name": "Andrew",
-      "email": "andrew@quartz.com",
-      "password": "password"
-    },
-    {
-      "name": "Joel",
-      "email": "joel@quartz.com",
-      "password": "password"
-    },
-    {
-      "name": "Itzel",
-      "email": "itzel@quartz.com",
-      "password": "password"
-    }
-  ]
-
