--- controllers/api/index.js
@@ -1,8 +1,0 @@
-const router = require('express').Router();
-const userRoutes = require('./userRoutes');
-const projectRoutes = require('./projectRoutes');
-
-router.use('/users', userRoutes);
-router.use('/projects', projectRoutes);
-
-module.exports = router;
