--- controllers/api/projectRoutes.js
@@ -1,49 +1,0 @@
-const router = require('express').Router();
-const { Project } = require('../../models');
-const withAuth = require('../../utils/auth');
-
-router.get('/', withAuth, async (req, res) => {
-  try {
-    Project.findAll().then((data) => {
-      res.status(200).json(data)
-    })
-  } catch (error) {
-    console.error('There Was An Error', error)
-    res.sendFile(path.join(__dirname, '../../public/404image.jpg'))
-  }
-})
-
-router.post('/', withAuth, async (req, res) => {
-  try {
-    const newProject = await Project.create({
-      ...req.body,
-      user_id: req.session.user_id,
-    });
-
-    res.status(200).json(newProject);
-  } catch (err) {
-    res.status(400).json(err);
-  }
-});
-
-router.delete('/:id', withAuth, async (req, res) => {
-  try {
-    const projectData = await Project.destroy({
-      where: {
-        id: req.params.id,
-        user_id: req.session.user_id,
-      },
-    });
-
-    if (!projectData) {
-      res.status(404).json({ message: 'No project found with this id!' });
-      return;
-    }
-
-    res.status(200).json(projectData);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-module.exports = router;
